refactor: update types architeture
